+++ BE/src/main/java/ifive/idrop/driver/service/DriverService.java
@@ -8,8 +8,8 @@ import ifive.idrop.entity.*;
 import ifive.idrop.entity.enums.PickUpStatus;
 import ifive.idrop.common.exception.CommonException;
 import ifive.idrop.common.exception.ErrorCode;
-import ifive.idrop.fcm.AlarmMessage;
-import ifive.idrop.fcm.NotificationUtill;
+import ifive.idrop.notification.AlarmMessage;
+import ifive.idrop.notification.NotificationUtill;
 import ifive.idrop.driver.repository.DriverRepository;
 import ifive.idrop.notification.domain.Notification;
 import ifive.idrop.parent.domain.Parent;
