+++ BE/src/main/java/ifive/idrop/notification/AlarmMessage.java
@@ -1,4 +1,4 @@
-package ifive.idrop.fcm;
+package ifive.idrop.notification;
 
 import lombok.Getter;
