+++ BE/src/main/java/ifive/idrop/parent/service/ParentService.java
@@ -13,8 +13,8 @@ import ifive.idrop.entity.*;
 import ifive.idrop.entity.enums.PickUpStatus;
 import ifive.idrop.common.exception.CommonException;
 import ifive.idrop.common.exception.ErrorCode;
-import ifive.idrop.fcm.AlarmMessage;
-import ifive.idrop.fcm.NotificationUtill;
+import ifive.idrop.notification.AlarmMessage;
+import ifive.idrop.notification.NotificationUtill;
 import ifive.idrop.driver.repository.DriverRepository;
 import ifive.idrop.parent.domain.Parent;
 import ifive.idrop.parent.repository.ParentRepository;
