+++ BE/src/main/java/ifive/idrop/pickup/service/PickUpService.java
@@ -4,8 +4,8 @@ import ifive.idrop.parent.domain.Parent;
 import ifive.idrop.pickup.domain.PickUp;
 import ifive.idrop.common.exception.CommonException;
 import ifive.idrop.common.exception.ErrorCode;
-import ifive.idrop.fcm.AlarmMessage;
-import ifive.idrop.fcm.NotificationUtill;
+import ifive.idrop.notification.AlarmMessage;
+import ifive.idrop.notification.NotificationUtill;
 import ifive.idrop.pickup.repository.PickUpRepository;
 import ifive.idrop.service.ImageService;
 import lombok.RequiredArgsConstructor;
